+++ lib/core.dart
@@ -19,6 +19,7 @@ library;
 
 export 'src/components/seo_component_format.dart';
 export 'src/components/seo_components.dart';
+export 'src/components/seo_rich_text.dart';
 export 'src/meta/seo_meta.dart';
 export 'src/meta/seo_schema.dart';
 export 'src/renderer/html_renderer.dart';